refactor: shared utils for temperature (#23)
- Move convert_temperature() and INPUT_REGISTER_*_TEMP constants to utils/temperature.py
- Update sensor/temperature.py and number/temperature.py to import from shared module
- Add native_unit_of_measurement to NumberEntityDescription (declarative)
- Add explicit __init__ to AirfiTemperatureNumber for consistency
- Remove native_unit_of_measurement property override
- Update tests to import from utils/temperature
