+++ custom_components/airfi/sensor/temperature.py
@@ -2,34 +2,21 @@
 
 from __future__ import annotations
 
-import math
 from typing import TYPE_CHECKING
 
 from custom_components.airfi.entity import AirfiEntity
+from custom_components.airfi.utils.temperature import (
+    INPUT_REGISTER_EXHAUST_AIR_TEMP,
+    INPUT_REGISTER_EXTRACT_AIR_TEMP,
+    INPUT_REGISTER_OUTDOOR_AIR_TEMP,
+    INPUT_REGISTER_SUPPLY_AIR_TEMP,
+    convert_temperature,
+)
 from homeassistant.components.sensor import SensorDeviceClass, SensorEntity, SensorEntityDescription, SensorStateClass
 from homeassistant.const import UnitOfTemperature
 
 if TYPE_CHECKING:
     from custom_components.airfi.coordinator import AirfiDataUpdateCoordinator
-
-INPUT_REGISTER_OUTDOOR_AIR_TEMP = 4
-INPUT_REGISTER_EXTRACT_AIR_TEMP = 6
-INPUT_REGISTER_EXHAUST_AIR_TEMP = 7
-INPUT_REGISTER_SUPPLY_AIR_TEMP = 8
-
-
-def convert_temperature(value: float) -> float:
-    """Convert an Airfi temperature register value into Celsius."""
-    if not math.isfinite(value):
-        return 0.0
-
-    converted = value
-
-    # Keep parity with Homebridge conversion for signed 16-bit encoded values.
-    if converted > 62803:
-        converted = value - 65535
-
-    return round(converted / 10.0, 1)
 
 
 ENTITY_DESCRIPTIONS: tuple[SensorEntityDescription, ...] = (
